2023-06-01 00:30LT Bug fixes: dash in regExp and decipher logic change. Minor cosmetic changes.
When building regexp for checking for allowed characters, dash was acting like a range indicator inside square brackets. When put it in the end of sequence it does not counted as a special symbol but as a usual literal.

A fatal error infiltrated like a spy. Apparently it was working somehow in previous versions avoiding my testcases. Now decipher algorithm fixed as it should be.

Also added method and menu item to output allowed characters to terminal.

Signed-off-by: kottofey <romeokontakt@gmail.com>